# Conflicts:
#	packages/apps-config/src/api/spec/index.ts
#	packages/apps-config/src/ui/logos/index.ts
#	packages/apps/public/locales/en/apps-config.json
#	packages/page-accounts/package.json
#	packages/page-accounts/src/Accounts/Account.tsx
#	packages/page-accounts/src/Accounts/BannerExtension.tsx
#	packages/page-accounts/src/modals/Create.tsx
#	packages/react-components/src/AddressInfo.tsx
#	packages/react-components/src/Progress.tsx
#	tsconfig.json
#	yarn.lock

+++ packages/page-staking/src/Actions/partials/Nominate.tsx
@@ -1,6 +1,7 @@
-// Copyright 2017-2021 @polkadot/app-staking authors & contributors
+// Copyright 2017-2022 @polkadot/app-staking authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import type { u32 } from '@polkadot/types-codec';
 import type { SortedTargets } from '../../types';
 import type { NominateInfo } from './types';
 
@@ -55,7 +56,7 @@ function Nominate ({ className = '', controllerId, nominating, onChange, stashId
   }, [api, onChange, selected]);
 
   const maxNominations = api.consts.staking.maxNominations
-    ? api.consts.staking.maxNominations.toNumber()
+    ? (api.consts.staking.maxNominations as u32).toNumber()
     : MAX_NOMINATIONS;
 
   return (
@@ -66,12 +67,14 @@ function Nominate ({ className = '', controllerId, nominating, onChange, stashId
           stashId={stashId}
         />
       )}
-      <Modal.Columns hint={
-        <>
-          <p>{t<string>('Nominators can be selected manually from the list of all currently available validators.')}</p>
-          <p>{t<string>('Once transmitted the new selection will only take effect in 2 eras taking the new validator election cycle into account. Until then, the nominations will show as inactive.')}</p>
-        </>
-      }>
+      <Modal.Columns
+        hint={
+          <>
+            <p>{t<string>('Nominators can be selected manually from the list of all currently available validators.')}</p>
+            <p>{t<string>('Once transmitted the new selection will only take effect in 2 eras taking the new validator election cycle into account. Until then, the nominations will show as inactive.')}</p>
+          </>
+        }
+      >
         <InputAddressMulti
           available={available}
           availableLabel={t<string>('candidate accounts')}
